Restructure into a generic turn-based game server and add Skull (#4)
This change overhauls the entire codebase into a generic game server
that centers around the concept of _rooms_. Players connect to an
existing room, or create a new one, when they open a WebSocket
connection to the server.

- Rooms can have 15 players max (subject to change in the future)
- Each room can only have one _game_ active at a time (e.g. Bravewength
  or Skull)
- Rooms support text-based chat (voice chat is planned for the future)
- Individual game implementations are responsible for features like
  ending/restarting the game, i.e., resetting the game state;
  preferably they will use terminology like "rounds" to avoid confusion
  with changing the type of game the room is playing
- Game implementations only need to satisfy a pretty simple interface
  where they can handle binary request packets and then emit binary
  state update packets to clients; this means some games can just use
  JSON-based protocols while others could optimize for maximum
  efficiency

Additionally, this change adds support for Skull, which is
alternatively referred to as "Skull and Roses". Having 2 game
implementations under my belt helped me settle on decent library
interface for the time being.